implement daily data import

+++ src/main/kotlin/at/rueckgr/kotlin/rocketbot/util/UserConfiguration.kt
@@ -36,7 +36,7 @@ data class TimePluginConfiguration(
 )
 
 data class SoccerPluginConfiguration(
-    val leagueId: Int?,
+    val leagueId: Long?,
     val season: Int?,
     val rounds: List<String>?,
     val apiKey: String?,
